Fix compilation errors for int64_t conversion
Build system and source fixes:
- Updated CMakeLists.txt: Added FitnessTests, JSComparisonTests, StepVerificationTests, DiagnosticTool
- Fixed Polygon.h: Changed polygonArea() forward declaration from double to int64_t
- Fixed GeometryUtil.cpp: Removed OrbitalTypes.h include, implemented normalizeVector() for int64_t, fixed noFitPolygonRectangle() to use CoordType
- Fixed PolygonExtractor.cpp: Added 'using namespace deepnest', fixed CoordType declarations, replaced noFitPolygon with MinkowskiSum
- Fixed NFPValidationTests.cpp: Removed OrbitalTypes.h include

Test results:
- GeometryValidationTests: 20/29 tests pass (69%)
- Critical tests passing: polygonArea(), ConvexHull, pointInPolygon(), getPolygonBounds()
- Some tolerance-related failures expected due to int64_t precision changes

+++ deepnest-cpp/include/deepnest/core/Polygon.h
@@ -12,7 +12,7 @@ namespace deepnest {
 
 // Forward declarations
 namespace GeometryUtil {
-    double polygonArea(const std::vector<Point>& polygon);
+    int64_t polygonArea(const std::vector<Point>& polygon);
     BoundingBox getPolygonBounds(const std::vector<Point>& polygon);
 }
